Add files via upload
This project is a comprehensive exploratory data analysis (EDA) on Netflix’s global catalog of Movies and TV Shows. It aims to uncover patterns, trends, and hidden insights in the content available on the platform using the power of Python’s data science stack. With over 8,800+ records and multiple mixed-type features such as text, dates, categories, and durations, this dataset provides the perfect opportunity to simulate real-world data analysis challenges.